Add communication features and refactor codebase
Created and incorporated several communication-related files and attributes to improve packet handling. In addition, the codebase was refactored, moving several methods and classes to improve organization and maintainability. This reorganization included the creation of new extension methods, removal of the "TalaryonExtensions" file, and introduction of asynchronous binary reader and writer classes. This work also involved minor updates to the Toolbox solution file.